+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Changed
 - start migrating to jsonrpc api
 - Update translations (22.09.2022)
+- remember last path in "save as" dialog
 
 ## Fixed
 - allow scanning of certain qr code types on welcome screen (account, url and text)